conv {kSamples}	R Documentation

Convolution of Two Discrete Distributions

Description

This function convolutes two discrete distribution, each given by strictly increasing support vectors and corresponding probability vectors.

Usage

	conv(x1,p1,x2,p2)

Arguments

x1
support vector of the first distribution, with strictly increasing elements.

p1
vector of probabilities corresponding to x1.

x2
support vector of the second distribution, with strictly increasing elements.

p2
vector of probabilities corresponding to x2.

Details

The convolution is performed in C, looping through all paired sums, augmenting existing values or inserting them with an update of the corresponding probabilities.

Value

A matrix with first column the new support vector and the second column the corresponding probability vector.

Examples

x1 <- c(1,2,3.5)
p1 <- c(.2,.3,.5)
x2 <- c(0,2.3,3,4)
p2 <- c(.1,.3,.3,.3)

conv(x1,p1,x2,p2)

[Package kSamples version 1.2-12 Index]